Moved the "parseSBSMessage()" method into SBSDecoder.
Parameters that are missing in Aircraft objects created from SBS-1 messages are filled in from the corresponding Aircraft object in the ArrayList which is maintained in the app. SBS-1 messages which are somehow corrupted (i.e. don't have 22 fields like a valid SBS-1 message) aren't parsed to stop the app from crashing.